chore: refactor component, update scroll insets
