--- src/navigator/constants.ts
@@ -1,1 +1,0 @@
-export const BOTTOM_SHEET_DEFAULT_HANDLE_HEIGHT = 24
